+++ tests/Debug/Whoops/DebuggerTest.php
@@ -37,18 +37,6 @@ class DebuggerTest extends \PHPUnit_Framework_TestCase
     }
 
     /**
-     * Tests if the debugger's environment is equal to the specified environment.
-     *
-     * @return void
-     */
-    public function testSetEnvironmentMethod()
-    {
-        $this->debugger->setEnvironment($this->environment);
-
-        $this->assertEquals($this->environment, $this->debugger->getEnvironment());
-    }
-
-    /**
      * Tests if the specified handler is in the debugger's list of handlers.
      *
      * @return void